added half2 kernal read from global memmory

+++ playingGound/playingGound/utils.hpp
@@ -2,10 +2,10 @@ template<typename T>
 float matrixMultiplicationCPU(T* c, const T* a, const T* b, unsigned int N) {
     float sum;
     auto start_cpu_time = std::chrono::high_resolution_clock::now();
-    for (int row = 0; row < N; row++) {
-        for (int col = 0; col < N; col++) {
+    for (unsigned int row = 0; row < N; row++) {
+        for (unsigned int col = 0; col < N; col++) {
             sum = 0.f;
-            for (int n = 0; n < N; n++) {
+            for (unsigned int n = 0; n < N; n++) {
                 //sum += (float)a[row * N + n] * (float)b[n * N + col];
                 //assuming the matrix is transposed for better Coalescing
                 sum += (float)a[row * N + n] * (float)b[col * N + n];
@@ -14,7 +14,7 @@ float matrixMultiplicationCPU(T* c, const T* a, const T* b, unsigned int N) {
         }
     }
     auto end_cpu_time = std::chrono::high_resolution_clock::now();
-    float cpu_time = std::chrono::duration<double, std::milli>(end_cpu_time - start_cpu_time).count();
+    float cpu_time = (float)std::chrono::duration<double, std::milli>(end_cpu_time - start_cpu_time).count();
     return cpu_time;
 
 }
